# anyharness/sdk/generated/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AnyHarness Runtime API",
    "description": "Local sidecar runtime for managing coding agent installations, credentials, and ACP sessions.",
    "license": {
      "name": "AGPL-3.0",
      "identifier": "AGPL-3.0"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/repo-roots/{repo_root_id}": {
      "get": {
        "tags": [
          "repo-roots"
        ],
        "operationId": "get_repo_root",
        "parameters": [
          {
            "name": "repo_root_id",
            "in": "path",
            "description": "Repo root ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Repo root",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RepoRoot"
                }
              }
            }
          },
          "404": {
            "description": "Repo root not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/v1/repo-roots/{repo_root_id}/mobility/prepare-destination": {
      "post": {
        "tags": [
          "repo-roots"
        ],
        "operationId": "prepare_repo_root_mobility_destination",
        "parameters": [
          {
            "name": "repo_root_id",
            "in": "path",
            "description": "Repo root ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PrepareRepoRootMobilityDestinationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Prepared repo root mobility destination",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrepareRepoRootMobilityDestinationResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Repo root not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "PrepareRepoRootMobilityDestinationRequest": {
        "type": "object",
        "required": [
          "requestedBranch",
          "requestedBaseSha"
        ],
        "properties": {
          "preferredWorkspaceName": {
            "type": [
              "string",
              "null"
            ]
          },
          "requestedBaseSha": {
            "type": "string"
          },
          "requestedBranch": {
            "type": "string"
          }
        }
      },
      "PrepareRepoRootMobilityDestinationResponse": {
        "type": "object",
        "required": [
          "workspace"
        ],
        "properties": {
          "workspace": {
            "$ref": "#/components/schemas/Workspace"
          }
        }
      },
      "ProblemDetails": {
        "type": "object",
        "required": [
          "type",
          "title",
          "status"
        ],
        "properties": {
          "code": {
            "type": [
              "string",
              "null"
            ]
          },
          "detail": {
            "type": [
              "string",
              "null"
            ]
          },
          "instance": {
            "type": [
              "string",
              "null"
            ]
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "title": {
            "type": "string"
          },
          "type": {
            "type": "string"
          }
        }
      },
...
      "RepoRoot": {
        "type": "object",
        "required": [
          "id",
          "kind",
          "path",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "createdAt": {
            "type": "string"
          },
          "defaultBranch": {
            "type": [
              "string",
              "null"
            ]
          },
          "displayName": {
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "type": "string"
          },
          "kind": {
            "$ref": "#/components/schemas/RepoRootKind"
          },
          "path": {
            "type": "string"
          },
          "remoteOwner": {
            "type": [
              "string",
              "null"
            ]
          },
          "remoteProvider": {
            "type": [
              "string",
              "null"
            ]
          },
          "remoteRepoName": {
            "type": [
              "string",
              "null"
            ]
          },
          "remoteUrl": {
            "type": [
              "string",
              "null"
            ]
          },
          "updatedAt": {
            "type": "string"
          }
        }
      },
      "RepoRootKind": {
        "type": "string",
        "enum": [
          "external",
          "managed"
        ]
      },
...
      "Workspace": {
        "type": "object",
        "required": [
          "id",
          "kind",
          "repoRootId",
          "path",
          "surface",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "createdAt": {
            "type": "string"
          },
          "currentBranch": {
            "type": [
              "string",
              "null"
            ]
          },
          "displayName": {
            "type": [
              "string",
              "null"
            ]
          },
          "executionSummary": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/WorkspaceExecutionSummary"
              }
            ]
          },
          "id": {
            "type": "string"
          },
          "kind": {
            "$ref": "#/components/schemas/WorkspaceKind"
          },
          "originalBranch": {
            "type": [
              "string",
              "null"
            ]
          },
          "path": {
            "type": "string"
          },
          "repoRootId": {
            "type": "string"
          },
          "surface": {
            "$ref": "#/components/schemas/WorkspaceSurface"
          },
          "updatedAt": {
            "type": "string"
          }
        }
      },
      "WorkspaceExecutionPhase": {
        "type": "string",
        "enum": [
          "running",
          "awaiting_permission",
          "idle",
          "errored"
        ]
      },
      "WorkspaceExecutionSummary": {
        "type": "object",
        "required": [
          "phase",
          "totalSessionCount",
          "liveSessionCount",
          "runningCount",
          "awaitingPermissionCount",
          "idleCount",
          "erroredCount"
        ],
        "properties": {
          "awaitingPermissionCount": {
            "type": "integer",
            "minimum": 0
          },
          "erroredCount": {
            "type": "integer",
            "minimum": 0
          },
          "idleCount": {
            "type": "integer",
            "minimum": 0
          },
          "liveSessionCount": {
            "type": "integer",
            "minimum": 0
          },
          "phase": {
            "$ref": "#/components/schemas/WorkspaceExecutionPhase"
          },
          "runningCount": {
            "type": "integer",
            "minimum": 0
          },
          "totalSessionCount": {
            "type": "integer",
            "minimum": 0
          }
        }
      },
      "WorkspaceKind": {
        "type": "string",
        "enum": [
          "worktree",
          "local"
        ]
      },
...
      "WorkspaceSurface": {
        "type": "string",
        "enum": [
          "standard",
          "cowork"
        ]
      }
    }
  }
}